KP-4479 Refactor verification data on preloading. fixed not working verifications.

+++ src/Core/BoundedContexts/Headquarters/WB.Core.BoundedContexts.Headquarters/AssignmentImport/Verifier/AssignmentRow.cs
@@ -45,7 +45,7 @@ namespace WB.Core.BoundedContexts.Headquarters.AssignmentImport.Verifier
 
     public class AssignmentRosterInstanceCode : AssignmentAnswer
     {
-        public int? Code { get; set; }
+        public decimal? Code { get; set; }
     }
 
     public class AssignmentAnswers: AssignmentValue
